Release

+++ src/methods/dialogs/getDialogs.ts
@@ -90,29 +90,23 @@ export const getDialogs = async (client: any, account: Account) => {
       } = dialogDb || {};
 
       if (blocked) {
-        try {
-          await client.invoke(
-            new GramJs.contacts.Block({
-              id: new GramJs.InputPeerUser({
-                userId: BigInt(user.id),
-                accessHash: BigInt(user.accessHash),
-              }),
-            })
-          );
-          await client.invoke(
-            new GramJs.messages.DeleteHistory({
-              peer: new GramJs.InputPeerUser({
-                userId: BigInt(user.id),
-                accessHash: BigInt(user.accessHash),
-              }),
-              revoke: true,
-            })
-          );
-        } catch (e: any) {
-          await sendToBot(
-            `Errorr 2: ${e.message}; User: ${JSON.stringify(user)}`
-          );
-        }
+        await client.invoke(
+          new GramJs.contacts.Block({
+            id: new GramJs.InputPeerUser({
+              userId: BigInt(user.id),
+              accessHash: BigInt(user.accessHash),
+            }),
+          })
+        );
+        await client.invoke(
+          new GramJs.messages.DeleteHistory({
+            peer: new GramJs.InputPeerUser({
+              userId: BigInt(user.id),
+              accessHash: BigInt(user.accessHash),
+            }),
+            revoke: true,
+          })
+        );
         continue;
       }
 
@@ -188,30 +182,32 @@ export const getDialogs = async (client: any, account: Account) => {
     } else {
       await saveBlockedRecipient(account.accountId, dialogId);
 
-      if (user && user.id && user.accessHash) {
-        try {
-          await client.invoke(
-            new GramJs.contacts.Block({
-              id: new GramJs.InputPeerUser({
-                userId: BigInt(user.id),
-                accessHash: BigInt(user.accessHash),
-              }),
-            })
-          );
-          await client.invoke(
-            new GramJs.messages.DeleteHistory({
-              peer: new GramJs.InputPeerUser({
-                userId: BigInt(user.id),
-                accessHash: BigInt(user.accessHash),
-              }),
-              revoke: true,
-            })
-          );
-        } catch (e: any) {
-          await sendToBot(
-            `Errorr: ${e.message}; User: ${JSON.stringify(user)}`
-          );
-        }
+      if (
+        user &&
+        user.id &&
+        user.accessHash &&
+        !user.deleted &&
+        !user.bot &&
+        !user.support &&
+        !user.self
+      ) {
+        await client.invoke(
+          new GramJs.contacts.Block({
+            id: new GramJs.InputPeerUser({
+              userId: BigInt(user.id),
+              accessHash: BigInt(user.accessHash),
+            }),
+          })
+        );
+        await client.invoke(
+          new GramJs.messages.DeleteHistory({
+            peer: new GramJs.InputPeerUser({
+              userId: BigInt(user.id),
+              accessHash: BigInt(user.accessHash),
+            }),
+            revoke: true,
+          })
+        );
       }
     }
   }
